Moved the test to unit test
